Flip the layout to put the vowels on the right hand.
This solves some problems with the initial layout, notably that the common
bigram 've' required alt fingering on the left hand with a matrix layout, and
was otherwise an uncomfortable stretch from a central position.  Additionally, a lot of common
shortcuts were missing from the left hand but are now accessible: t, f, c were
the most obvious ones that were originally missing.