Move b down to from QWERTY u to QWERTY h.
This improves the comfort of the combination "be", in particular for staggered
keyboards.  The backspace which was on that key has now been moved to QWERTY n,
and period/comma to QWERTY u (which I'm less happy about but anticipate
backspace being the more frequently used key and n is better placed so having
period a bit out of the way is a little compromise).

Note this makes the regex combination .* a right index pull on the symbol layer.